+++ Cybereason/cybereason_modules/connector_pull_events.py
@@ -1,5 +1,6 @@
 import signal
 import time
+from datetime import datetime, timedelta
 from collections import defaultdict
 from collections.abc import Generator
 from functools import cached_property
@@ -9,7 +10,9 @@ from posixpath import join as urljoin
 
 import orjson
 import requests
+from cachetools import Cache, LRUCache
 from sekoia_automation.connector import Connector, DefaultConnectorConfiguration
+from sekoia_automation.checkpoint import CheckpointTimestamp, TimeUnit
 
 from cybereason_modules import CybereasonModule
 from cybereason_modules.client import ApiClient
@@ -45,12 +48,36 @@ class CybereasonEventConnector(Connector):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.from_date: int = (int(time.time()) - 60) * 1000  # milliseconds
+
+        self.cursor = CheckpointTimestamp(
+            time_unit=TimeUnit.MILLISECOND,
+            path=self._data_path,
+            start_at=timedelta(days=90),
+            ignore_older_than=timedelta(days=90),
+        )
+
+        self.from_date: int = self.cursor.offset
+        self.events_cache: Cache = self.load_events_cache()
         self._stop_event = Event()  # Event to notify we must stop the thread
 
         # Register signal to terminate thread
         signal.signal(signal.SIGINT, self.exit)
         signal.signal(signal.SIGTERM, self.exit)
+
+    def load_events_cache(self) -> Cache:
+        result: LRUCache = LRUCache(maxsize=1000)
+
+        with self.cursor._context as cache:
+            events_ids = cache.get("events_cache", [])
+
+        for event_id in events_ids:
+            result[event_id] = 1
+
+        return result
+
+    def save_events_cache(self, events_cache: Cache) -> None:
+        with self.cursor._context as cache:
+            cache["events_cache"] = list(events_cache.keys())
 
     @cached_property
     def client(self):
@@ -265,15 +292,8 @@ class CybereasonEventConnector(Connector):
         Fetch the last malops from the Cybereason API
         """
         from_date = self.from_date
-        now = int(time.time()) * 1000  # milliseconds
-
-        # We don't retrieve messages older than one hour
-        one_hour_ago = now - 3600000
-        if from_date < one_hour_ago:
-            from_date = one_hour_ago
-
         # compute the ending time to retrieve malops (Currently, now)
-        to_date = now
+        to_date = int(time.time()) * 1000
 
         # fetch malops for the timerange
         next_malops = self.fetch_malops(from_date, to_date)
@@ -281,12 +301,14 @@ class CybereasonEventConnector(Connector):
 
         most_recent_date_seen = from_date
         for malop in next_malops:
+
+            malop_uuid = malop["guid"]
+            self.events_cache[malop_uuid] = 1
+
             # save the greater date ever seen
             event_date = int(malop["lastUpdateTime"])
             if event_date > most_recent_date_seen:
-                most_recent_date_seen = (
-                    event_date + 1
-                )  # add 1 milli-seconds to exclude the current malop from the next search
+                most_recent_date_seen = event_date
 
             # check if the malop is an AI Hunt malop (EDR) or a generic one
             is_edr = malop.get("edr", False)
@@ -313,8 +335,11 @@ class CybereasonEventConnector(Connector):
             else:
                 yield from self.enrich_generic_malop(malop)
 
+        self.save_events_cache(self.events_cache)
+
         # save the most recent date and compute the lag
         if most_recent_date_seen > self.from_date:
+            self.cursor.offset = most_recent_date_seen
             self.from_date = most_recent_date_seen
             current_lag = int(time.time() - (most_recent_date_seen / 1000))
             EVENTS_LAG.labels(intake_key=self.configuration.intake_key).set(current_lag)